fov init points calculated

+++ src/danae/dn_put_pixel.c
@@ -6,13 +6,13 @@
 /*   By: becastro <becastro@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/13 13:12:11 by becastro          #+#    #+#             */
-/*   Updated: 2022/12/15 15:37:32 by becastro         ###   ########.fr       */
+/*   Updated: 2022/12/16 15:07:33 by becastro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "danae.h"
 
-void	dn_put_pixel(t_mlx *mlx, uint16_t x, uint16_t y, int color)
+void	dn_put_pixel(t_mlx *mlx, int x, int y, int color)
 {
 	char	*dst;
